Separated code for better organization
Separated frontend code into separate files for better organization
