+++ frontend/src/App.js
@@ -1,7 +1,9 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes, Link } from "react-router-dom";
-import Form from "./form";
 import "./App.css";
+import Welcome from "./components/welcome";
+import Success from "./components/success";
+import Form from "./components/form";
 
 const App = () => {
   return (
@@ -12,31 +14,6 @@ const App = () => {
         <Route path="/success" element={<Success />} />
       </Routes>
     </Router>
-  );
-};
-
-// Welcome component for the "/" route
-const Welcome = () => {
-  return (
-    <div className="welcome-container">
-      <h1>Welcome to Patient Submission Portal</h1>
-      <p>Click the button below to submit patient information.</p>
-      <Link to="/submission">
-        <button className="button">Go to Submission Form</button>
-      </Link>
-    </div>
-  );
-};
-
-const Success = () => {
-  return (
-    <div className="welcome-container">
-      <h1>Submission Successful!</h1>
-      <p>Thank you for submitting the patient information.</p>
-      <Link to="/">
-        <button className="button">Go Back to Home</button>
-      </Link>
-    </div>
   );
 };
